# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finance Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  income: "income",
  expense: "expense",
} as const;

export interface Transaction {
  id: number;
  description: string;
  amount: number;
  type: TransactionType;
  category: string;
  date: string;
  createdAt: string;
  updatedAt: string;
}

export type CreateTransactionBodyType =
  (typeof CreateTransactionBodyType)[keyof typeof CreateTransactionBodyType];

export const CreateTransactionBodyType = {
  income: "income",
  expense: "expense",
} as const;

export interface CreateTransactionBody {
  description: string;
  amount: number;
  type: CreateTransactionBodyType;
  category: string;
  date: string;
}

export type UpdateTransactionBodyType =
  (typeof UpdateTransactionBodyType)[keyof typeof UpdateTransactionBodyType];

export const UpdateTransactionBodyType = {
  income: "income",
  expense: "expense",
} as const;

export interface UpdateTransactionBody {
  description?: string;
  amount?: number;
  type?: UpdateTransactionBodyType;
  category?: string;
  date?: string;
}

export interface DashboardSummary {
  totalBalance: number;
  totalIncome: number;
  totalExpenses: number;
  monthlySavings: number;
  monthlyIncome: number;
  monthlyExpenses: number;
  transactionCount: number;
}

export interface MonthlyTrend {
  month: string;
  income: number;
  expenses: number;
  savings: number;
}

export interface CategoryBreakdown {
  category: string;
  amount: number;
  percentage: number;
  count: number;
}

export interface Insights {
  highestSpendingCategory: string;
  highestSpendingAmount: number;
  savingsRate: number;
  monthOverMonthChange: number;
  topExpenseCategories: string[];
  averageTransactionAmount: number;
}

export type ListTransactionsParams = {
  /**
   * @nullable
   */
  category?: string | null;
  /**
   * @nullable
   */
  type?: string | null;
  /**
   * @nullable
   */
  search?: string | null;
  /**
   * @nullable
   */
  sortBy?: string | null;
  /**
   * @nullable
   */
  sortOrder?: string | null;
};