# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Finance Dashboard API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: transactions
    description: Transaction operations
  - name: summary
    description: Dashboard summary operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List all transactions
      parameters:
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: type
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: sortBy
          in: query
          schema:
            type: ["string", "null"]
        - name: sortOrder
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a new transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionBody"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Validation error

  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a transaction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
    patch:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTransactionBody"
      responses:
        "200":
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Validation error
        "404":
          description: Not found
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /summary:
    get:
      operationId: getDashboardSummary
      tags: [summary]
      summary: Get dashboard financial summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /summary/monthly:
    get:
      operationId: getMonthlyTrend
      tags: [summary]
      summary: Get monthly income vs expense trend
      responses:
        "200":
          description: Monthly trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyTrend"

  /summary/category-breakdown:
    get:
      operationId: getCategoryBreakdown
      tags: [summary]
      summary: Get spending breakdown by category
      responses:
        "200":
          description: Category breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryBreakdown"

  /summary/insights:
    get:
      operationId: getInsights
      tags: [summary]
      summary: Get financial insights
      responses:
        "200":
          description: Financial insights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Insights"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Transaction:
      type: object
      properties:
        id:
          type: integer
        description:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        date:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - description
        - amount
        - type
        - category
        - date
        - createdAt
        - updatedAt

    CreateTransactionBody:
      type: object
      properties:
        description:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        date:
          type: string
      required:
        - description
        - amount
        - type
        - category
        - date

    UpdateTransactionBody:
      type: object
      properties:
        description:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        date:
          type: string

    DashboardSummary:
      type: object
      properties:
        totalBalance:
          type: number
        totalIncome:
          type: number
        totalExpenses:
          type: number
        monthlySavings:
          type: number
        monthlyIncome:
          type: number
        monthlyExpenses:
          type: number
        transactionCount:
          type: integer
      required:
        - totalBalance
        - totalIncome
        - totalExpenses
        - monthlySavings
        - monthlyIncome
        - monthlyExpenses
        - transactionCount

    MonthlyTrend:
      type: object
      properties:
        month:
          type: string
        income:
          type: number
        expenses:
          type: number
        savings:
          type: number
      required:
        - month
        - income
        - expenses
        - savings

    CategoryBreakdown:
      type: object
      properties:
        category:
          type: string
        amount:
          type: number
        percentage:
          type: number
        count:
          type: integer
      required:
        - category
        - amount
        - percentage
        - count

    Insights:
      type: object
      properties:
        highestSpendingCategory:
          type: string
        highestSpendingAmount:
          type: number
        savingsRate:
          type: number
        monthOverMonthChange:
          type: number
        topExpenseCategories:
          type: array
          items:
            type: string
        averageTransactionAmount:
          type: number
      required:
        - highestSpendingCategory
        - highestSpendingAmount
        - savingsRate
        - monthOverMonthChange
        - topExpenseCategories
        - averageTransactionAmount