# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finance Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all transactions
 */
export const ListTransactionsQueryParams = zod.object({
  category: zod.coerce.string().nullish(),
  type: zod.coerce.string().nullish(),
  search: zod.coerce.string().nullish(),
  sortBy: zod.coerce.string().nullish(),
  sortOrder: zod.coerce.string().nullish(),
});

export const ListTransactionsResponseItem = zod.object({
  id: zod.number(),
  description: zod.string(),
  amount: zod.number(),
  type: zod.enum(["income", "expense"]),
  category: zod.string(),
  date: zod.string(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListTransactionsResponse = zod.array(ListTransactionsResponseItem);

/**
 * @summary Create a new transaction
 */
export const CreateTransactionBody = zod.object({
  description: zod.string(),
  amount: zod.number(),
  type: zod.enum(["income", "expense"]),
  category: zod.string(),
  date: zod.string(),
});

/**
 * @summary Get a transaction by ID
 */
export const GetTransactionParams = zod.object({
  id: zod.coerce.number(),
});

export const GetTransactionResponse = zod.object({
  id: zod.number(),
  description: zod.string(),
  amount: zod.number(),
  type: zod.enum(["income", "expense"]),
  category: zod.string(),
  date: zod.string(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update a transaction
 */
export const UpdateTransactionParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateTransactionBody = zod.object({
  description: zod.string().optional(),
  amount: zod.number().optional(),
  type: zod.enum(["income", "expense"]).optional(),
  category: zod.string().optional(),
  date: zod.string().optional(),
});

export const UpdateTransactionResponse = zod.object({
  id: zod.number(),
  description: zod.string(),
  amount: zod.number(),
  type: zod.enum(["income", "expense"]),
  category: zod.string(),
  date: zod.string(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a transaction
 */
export const DeleteTransactionParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get dashboard financial summary
 */
export const GetDashboardSummaryResponse = zod.object({
  totalBalance: zod.number(),
  totalIncome: zod.number(),
  totalExpenses: zod.number(),
  monthlySavings: zod.number(),
  monthlyIncome: zod.number(),
  monthlyExpenses: zod.number(),
  transactionCount: zod.number(),
});

/**
 * @summary Get monthly income vs expense trend
 */
export const GetMonthlyTrendResponseItem = zod.object({
  month: zod.string(),
  income: zod.number(),
  expenses: zod.number(),
  savings: zod.number(),
});
export const GetMonthlyTrendResponse = zod.array(GetMonthlyTrendResponseItem);

/**
 * @summary Get spending breakdown by category
 */
export const GetCategoryBreakdownResponseItem = zod.object({
  category: zod.string(),
  amount: zod.number(),
  percentage: zod.number(),
  count: zod.number(),
});
export const GetCategoryBreakdownResponse = zod.array(
  GetCategoryBreakdownResponseItem,
);

/**
 * @summary Get financial insights
 */
export const GetInsightsResponse = zod.object({
  highestSpendingCategory: zod.string(),
  highestSpendingAmount: zod.number(),
  savingsRate: zod.number(),
  monthOverMonthChange: zod.number(),
  topExpenseCategories: zod.array(zod.string()),
  averageTransactionAmount: zod.number(),
});