# lib/api-zod/src/generated/types/insights.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finance Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface Insights {
  highestSpendingCategory: string;
  highestSpendingAmount: number;
  savingsRate: number;
  monthOverMonthChange: number;
  topExpenseCategories: string[];
  averageTransactionAmount: number;
}