# lib/api-zod/src/generated/types/monthlyTrend.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finance Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface MonthlyTrend {
  month: string;
  income: number;
  expenses: number;
  savings: number;
}