# lib/api-zod/src/generated/types/transaction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finance Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionType } from "./transactionType";

export interface Transaction {
  id: number;
  description: string;
  amount: number;
  type: TransactionType;
  category: string;
  date: string;
  createdAt: string;
  updatedAt: string;
}