# lib/api-zod/src/generated/types/updateTransactionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finance Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTransactionBodyType } from "./updateTransactionBodyType";

export interface UpdateTransactionBody {
  description?: string;
  amount?: number;
  type?: UpdateTransactionBodyType;
  category?: string;
  date?: string;
}